refactor(formatNotif): move formatNotification util in /util folder
