+++ packages/web-app/src/util/formatNotification.js
